refactor(sonar-idl): flatten module structure and add variable-length discriminator support
## Changes

### Module Structure (Breaking)
- Flatten `models/` (4 files) into single `idl.rs`
- Flatten `parser/` (2 files) into `decode.rs` + `indexed.rs`
- Move tests to `tests/` directory
- Remove circular dependency between `raw.rs` and `indexed.rs`

### Discriminator Improvements
- **Instructions**: Support arbitrary-length discriminators with longest-first matching
- **Accounts**: Add `discriminator` field to `IdlTypeDefinition`, fallback to `sighash("account", name)`
- **Events**: Support arbitrary-length discriminators, fallback to `sighash("event", name)`

### API Changes
- Public API unchanged: `IndexedIdl`, `IdlParsedInstruction`, `IdlParsedField`, `is_cpi_event_data`, `sighash`
- Internal: All discriminators use `Vec<(Vec<u8>, usize)>` for variable-length support

### Quality
- All 540 workspace tests pass
- `cargo fmt` clean
- `cargo clippy -D warnings` clean
