--- crates/sonar-idl/src/models/mod.rs
@@ -1,18 +1,0 @@
-mod idl;
-mod raw;
-mod serde;
-mod types;
-
-pub(crate) use self::idl::{Idl, IdlAccountItem, IdlArg, IdlEvent, IdlInstruction, IdlMetadata};
-#[cfg(test)]
-pub(crate) use self::idl::{IdlAccount, IdlAccounts};
-pub(crate) use self::raw::RawAnchorIdl;
-pub(crate) use self::types::{
-    DefinedType, IdlArrayType, IdlField, IdlFields, IdlType, IdlTypeDefinition,
-    IdlTypeDefinitionKind,
-};
-
-pub(crate) use self::serde::deserialize_optional_idl_fields;
-
-#[cfg(test)]
-mod tests;
